+++ prost-reflect/src/lib.rs
@@ -109,8 +109,8 @@ mod dynamic;
 mod reflect;
 
 pub use self::descriptor::{
-    DescriptorError, EnumDescriptor, EnumValueDescriptor, FieldDescriptor, FileDescriptor, Kind,
-    MessageDescriptor, MethodDescriptor, OneofDescriptor, ServiceDescriptor,
+    DescriptorError, EnumDescriptor, EnumValueDescriptor, ExtensionDescriptor, FieldDescriptor,
+    FileDescriptor, Kind, MessageDescriptor, MethodDescriptor, OneofDescriptor, ServiceDescriptor,
 };
 pub use self::dynamic::{DynamicMessage, MapKey, Value};
 pub use self::reflect::ReflectMessage;